requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
discord.py==2.5.2
dotenv==0.9.9
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langsmith==0.3.43
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
orjson==3.10.18
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
zstandard==0.23.0
